Moved getInitialState and related logic to login Removed circular dependency on serverIO
