+++ ui/src/components/Login/Login.jsx
@@ -7,6 +7,7 @@ import logo from '../../img/mxcube_logo20.png';
 import loader from '../../img/loader.gif';
 import withRouter from '../WithRouter';
 import styles from './Login.module.css';
+import { serverIO } from '../../serverIO';
 
 function LoginComponent(props) {
   const { router, loading, logIn, showError, errorMessage } = props;
@@ -17,8 +18,9 @@ function LoginComponent(props) {
     formState: { errors },
   } = useForm({ defaultValues: { username: '', password: '' } });
 
-  function handleSubmit(data) {
-    logIn(data.username.toLowerCase(), data.password, router.navigate);
+  async function handleSubmit(data) {
+    await logIn(data.username.toLowerCase(), data.password, router.navigate);
+    serverIO.listen();
   }
 
   return (
